+++ base/cvd/MODULE.bazel
@@ -63,3 +63,25 @@ include("//cuttlefish/host/commands/append_squashfs_overlay:append_squashfs_over
 include("//cuttlefish/host/commands/vhost_user_input:vhost_user_input.MODULE.bazel")
 include("//cuttlefish/host/commands/vhost_user_media:vhost_user_media.MODULE.bazel")
 include("//toolchain:toolchain.MODULE.bazel")
+
+bazel_dep(name = "gazelle", version = "0.46.0")
+bazel_dep(name = "rules_go", version = "0.60.0")
+
+go_sdk = use_extension("@rules_go//go:extensions.bzl", "go_sdk")
+go_sdk.download(version = "1.24.12")
+
+go_deps = use_extension("@gazelle//:extensions.bzl", "go_deps")
+go_deps.from_file(go_mod = "//host_tests:go.mod")
+
+use_repo(
+    go_deps,
+    "com_github_google_go_cmp",
+)
+
+new_local_repository = use_repo_rule("@bazel_tools//tools/build_defs/repo:local.bzl", "new_local_repository")
+
+new_local_repository(
+    name = "cuttlefish_debian",
+    build_file_content = 'exports_files(glob(["**"]))\n',
+    path = "../debian",
+)
