+++ base/cvd/host_tests/common/BUILD.bazel
@@ -1,0 +1,37 @@
+# Copyright (C) 2026 The Android Open Source Project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "common",
+    srcs = [
+        "cvdalloc.go",
+        "dnsmasq.go",
+        "init_filesystem_ns.go",
+        "init_network_ns.go",
+        "ip.go",
+        "nft.go",
+        "sandbox.go",
+        "state.go",
+        "static_resources.go",
+    ],
+    embedsrcs = ["dnsmasq_shim.sh"],
+    importpath = "github.com/google/android-cuttlefish/base/cvd/host_tests/common",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_google_go_cmp//cmp",
+        "@rules_go//go/runfiles",
+    ],
+)
