+++ base/cvd/host_tests/common/static_resources.go
