+++ base/cvd/host_tests/go.mod
@@ -1,0 +1,7 @@
+module github.com/google/android-cuttlefish/base/cvd/host_tests
+
+go 1.24.0
+
+require github.com/google/go-cmp v0.6.0
+
+require github.com/bazelbuild/rules_go v0.60.0 // indirect
